requirements.txt
python-keycloak==3.9.1
pytest==8.3.5
pytest-mock==3.14.0